#34 - move handler into logic folder
