+++ server/src/logic/handler/routes/recipe.handler.ts
@@ -1,6 +1,6 @@
 import { Request, Response } from "express";
-import { RecipeService, UserService } from "./../../logic/services/index";
-import { Recipe } from "./../../models/index";
+import { RecipeService, UserService } from "./../../../logic/services/index";
+import { Recipe } from "./../../../models/index";
 
 export class RecipeHandler {
