+++ server/src/logic/handler/validation/model-validation.handler.ts
@@ -1,6 +1,6 @@
 import { NextFunction, Request, Response, Router } from "express";
-import { ModelValidationMiddleware } from "./../../logic/middleware/index";
-import { ApiError, DbError, Recipe, ValidationError } from "./../../models/index";
+import { ModelValidationMiddleware } from "./../../../logic/middleware/index";
+import { ApiError, DbError, Recipe, ValidationError } from "./../../../models/index";
 
 export class ModelValidationHandler {
