+++ server/src/routes/api/v1/users/users.route.ts
@@ -1,5 +1,5 @@
 import { Router } from "express";
-import { AuthValidationHandler, UserHandler } from "./../../../../handler/index";
+import { AuthValidationHandler, UserHandler } from "./../../../../logic/handler/index";
 import { UserService } from "./../../../../logic/services/user.service";
 
 export class UsersRoute {
