+++ server/test/recipe.route.test.ts
@@ -4,7 +4,7 @@ import * as chai from "chai";
 import * as express from "express";
 import * as mocha from "mocha";
 import * as sinon from "sinon";
-import { RecipeHandler } from "./../src/handler/index";
+import { RecipeHandler } from "./../src/logic/handler/index";
 import { RecipeService, UserService } from "./../src/logic/services/index";
 import { ApiError, DbError, ERRORS, Recipe } from "./../src/models/index";
 import Server from "./../src/server";
